Move validation shader validation
