+++ compositor_render/src/transformations/shader/validation/shader_header.wgsl
