+++ compositor_render/src/wgpu.rs
@@ -5,7 +5,6 @@ mod ctx;
 pub(crate) mod format;
 pub(crate) mod texture;
 pub(crate) mod utils;
-pub(crate) mod validation;
 
 pub(crate) use ctx::WgpuCtx;
