+++ compositor_render/src/wgpu/ctx.rs
@@ -50,9 +50,7 @@ impl WgpuCtx {
             None,
         ))?;
 
-        let shader_header =
-            naga::front::wgsl::parse_str(include_str!("./validation/shader_header.wgsl"))
-                .expect("failed to parse the shader header file");
+        let shader_header = crate::transformations::shader::validation::shader_header();
 
         let scope = WgpuErrorScope::push(&device);
